pyproject.toml
[project]
name = "strands-example"
version = "0.1.0"
description = "Agent with Strands and PostgreSQL integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[gemini]>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "strands-agents-tools>=0.2.16",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
